+++ WPILib/DigitalInput.cpp
@@ -26,13 +26,13 @@ void DigitalInput::InitDigitalInput(UINT8 moduleNumber, UINT32 channel)
     Resource::CreateResourceObject(&interruptsResource, tInterrupt::kNumSystems);
     if (!CheckDigitalModule(moduleNumber))
     {
-	snprintf(buf, 64, "Digital Module %d", moduleNumber);
+	snprintf(buf, 64, "Digital Module %u", moduleNumber);
 	wpi_setWPIErrorWithContext(ModuleIndexOutOfRange, buf);
 	return;
     }
     if (!CheckDigitalChannel(channel))
     {
-	snprintf(buf, 64, "Digital Channel %lu", channel);
+	snprintf(buf, 64, "Digital Channel %u", (u_int) channel);
 	wpi_setWPIErrorWithContext(ChannelIndexOutOfRange, buf);
 	return;
     }
